requirements.txt
Flask==2.3.2
python-dotenv==1.0.1
openai==1.31.0
pymupdf==1.23.1
pandas==2.1.1
matplotlib==3.8.0
